peptidemapper/requirements.txt
appdirs==1.4.0
beautifulsoup4==4.5.3
biopython==1.68
bioservices==1.4.14
colorama==0.3.7
diff-match-patch==20121119
dj-static==0.0.6
Django==1.8
django-crispy-forms==1.6.1
django-import-export==0.5.1
django-ipware==1.1.6
easydev==0.9.30
elasticsearch==6.1.1
gevent==1.2.0
goatools==0.7.9
greenlet==0.4.11
grequests==0.3.0
names==0.3.0
nose==1.3.7
numpy==1.11.3
pandas==0.20.3
patsy==0.4.1
pexpect==4.2.1
psutil==5.4.5
ptyprocess==0.5.1
python-dateutil==2.6.1
pytz==2017.2
requests==2.12.4
requests-cache==0.4.13
scipy==0.19.1
selenium==3.0.1
six==1.11.0
static3==0.7.0
statsmodels==0.8.0
suds-jurko==0.6
tablib==0.11.2
urllib3==1.22
wget==3.2
wrapt==1.10.8
xlrd==1.1.0
XlsxWriter==1.0.0
xmltodict==0.10.2
